finish mapr endering while building new map
some notes, water has to have at least 4 tiles to runder correctly it
will be like 2x2 or two rows after each other or two column after each
other. we changed the edges and we will discuss this with Nitta tomorrow
to see if this is the way he want it to be. other than that everything
should be fine